Add contains function
Use contains() to test if a list contains a defined value

The contains function works by attempting to find the index of a value within a list.
If it is unable to do so, the index function will return a false.
If it does return a value, that value is then converted in to a `true` via `not not`.